Revert "8-byte long is replaced with 4-byte int"
This reverts commit 1b9bb158bdd03be30e082bedeb2c57ebab77d13f.

+++ test/srcGMRESPeriodicBCTest/main.cpp
@@ -475,7 +475,7 @@ int main(int argc,char **argv) {
 
   //if the new mesh was generated => rename created mesh.msh into amr.sig=0x%lx.mesh.bin
   if (NewMeshGeneratedFlag==true) {
-    unsigned MeshSignature=PIC::Mesh::mesh->getMeshSignature();
+    unsigned long MeshSignature=PIC::Mesh::mesh->getMeshSignature();
 
     if (PIC::Mesh::mesh->ThisThread==0) {
       char command[300];
@@ -516,7 +516,7 @@ int main(int argc,char **argv) {
 
   int parSize=10;
   cTreeNodeAMR<PIC::Mesh::cDataBlockAMR>* newNode;
-  int newParticle;
+  long int newParticle;
 
   if (PIC::ThisThread==0) printf("test2\n");
